my orignial machine repurposed
Since the first machine's code was not clean, I recreated a new and clean machine and updated the code.  I think it works now.  Thank you to Alfred B for helping me sort this one out.  I am so glad I persisted with fixing it as I rather like it, as it is so quirky .